pacman has been temporarily removed
Sorry for the title on my last commit, I made it on mobile on the github site, and I hit enter instead of backspace. I removed pacman because I was having a bunch of issues with it. I don’t know when it will be back. I left most of the code in to make it easier to readd pacman. I just made it so the user could not select Pac-Man.